requirements.txt
Pillow==9.0.1
urllib3==1.26.8
